% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inferDriverDominceScore.R
\name{inferDriverDominanceScore}
\alias{inferDriverDominanceScore}
\title{Infer the Driver Dominances Scores for Patients and Genes.}
\usage{
inferDriverDominanceScore(
  maf,
  tsb = NULL,
  driver = NULL,
  topGenes = 10,
  vafCol = NULL,
  segFile = NULL,
  ignChr = NULL,
  minVaf = 0,
  maxVaf = 1,
  useSyn = FALSE
)
}
\arguments{
\item{maf}{an \code{\link{MAF}} object generated by \code{\link{read.maf}}}

\item{tsb}{specify sample names (Tumor_Sample_Barcodes) for which clustering has to be done.}

\item{driver}{The driver genes list(Hugo_Symbol), defaults setting the \strong{topGenes}}

\item{topGenes}{if \code{driver} is NULL, uses top n number of most frequently mutated genes. Defaults to 10.}

\item{vafCol}{manually specify column name for vafs. Default looks for column 't_vaf'}

\item{segFile}{path to CBS segmented copy number file. Column names should be Sample, Chromosome, Start, End, Num_Probes and Segment_Mean (log2 scale).}

\item{ignChr}{ignore these chromosomes from analysis. e.g, sex chromsomes chrX, chrY. Default NULL.}

\item{minVaf}{filter low frequency variants. Low vaf variants maybe due to sequencing error. Default 0. (on the scale of 0 to 1)}

\item{maxVaf}{filter high frequency variants. High vaf variants maybe due to copy number alterations or impure tumor. Default 1. (on the scale of 0 to 1)}

\item{useSyn}{Use synonymous variants. Default FALSE.}
}
\description{
Infer the Driver Dominances Scores for Patients and Genes.
}
